services/calc_service/requirements.txt
Flask==3.0.3
gunicorn==22.0.0
Pillow==10.4.0
imagehash==4.3.1
matplotlib==3.9.0
numpy==1.26.4
seaborn==0.13.2
